# specs/005-swagger-api-docs/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: "API-X"
  version: "0.1.0"
  description: |
    **API-X** adalah backend REST API untuk manajemen event dan pendaftaran peserta,
    dilengkapi dengan autentikasi berbasis JWT dan laporan statistik untuk admin.

    ## Cara Menggunakan API Ini

    Sebagian besar endpoint memerlukan autentikasi. Ikuti langkah berikut:

    1. **Daftar akun** — Panggil `POST /auth/register` dengan nama, email, dan kata sandi.
    2. **Login** — Panggil `POST /auth/login` dengan email dan kata sandi untuk mendapatkan JWT access token.
    3. **Salin token** — Dari respons login, salin nilai `access_token`.
    4. **Masukkan token** — Klik tombol **Authorize** di atas, lalu masukkan token dalam format `Bearer <token>`.
    5. **Akses endpoint** — Semua endpoint terproteksi kini dapat digunakan langsung dari halaman ini.

    ## Peran Pengguna

    | Peran | Keterangan |
    |-------|-----------|
    | **Publik** | Dapat diakses tanpa autentikasi |
    | **Memerlukan token** | Memerlukan JWT token pengguna terdaftar |
    | **Khusus admin** | Memerlukan JWT token dengan role `admin` |

    ## Format Error

    Semua error menggunakan struktur JSON yang konsisten:

    ```json
    {
      "error": {
        "code": "KODE_ERROR",
        "message": "Pesan yang dapat dibaca manusia.",
        "httpStatus": 4xx
      }
    }
    ```
servers:
  - url: "http://localhost:8000"
    description: "Lokal (development)"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token yang diperoleh dari `POST /auth/login`.

        **Cara mendapatkan token**:
        1. Panggil `POST /auth/login` dengan email dan kata sandi yang valid.
        2. Salin nilai `access_token` dari respons.
        3. Klik tombol **Authorize** di atas dan masukkan nilai token (tanpa awalan "Bearer").

        **Format header**: `Authorization: Bearer <token>`

        Token berlaku selama **60 menit** sejak diterbitkan (HS256 JWT).
  schemas:
...
    EventCreate:
      type: object
      required: [title, description, date, registration_deadline, quota]
      description: "Data yang diperlukan untuk membuat event baru. Hanya dapat dilakukan oleh admin."
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: "Judul event yang akan ditampilkan kepada peserta."
          examples:
            - "Workshop Python Lanjutan"
        description:
          type: string
          minLength: 1
          description: "Deskripsi lengkap event, termasuk agenda dan informasi pembicara."
          examples:
            - "Workshop intensif satu hari tentang FastAPI dan Pydantic v2. Cocok untuk developer yang sudah familiar dengan Python dasar."
        date:
          type: string
          format: date-time
          description: "Tanggal dan waktu pelaksanaan event (ISO 8601 dengan informasi timezone). Harus di masa depan saat event dibuat."
          examples:
            - "2026-06-15T09:00:00+07:00"
        registration_deadline:
          type: string
          format: date-time
          description: "Batas waktu pendaftaran (ISO 8601 dengan timezone). Harus sebelum atau sama dengan `date`."
          examples:
            - "2026-06-10T23:59:59+07:00"
        quota:
          type: integer
          minimum: 1
          description: "Kapasitas maksimum peserta. Harus minimal 1."
          examples:
            - 50
    EventUpdate:
      type: object
      description: "Data untuk memperbarui event yang sudah ada. Semua field bersifat opsional — hanya field yang disertakan yang akan diperbarui (partial update). Khusus admin."
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: "Judul event baru (opsional)."
          examples:
            - "Workshop Python Lanjutan — Edisi Revisi"
        description:
          type: string
          minLength: 1
          description: "Deskripsi event baru (opsional)."
          examples:
            - "Deskripsi yang diperbarui."
        date:
          type: string
          format: date-time
          description: "Tanggal dan waktu pelaksanaan baru (opsional)."
          examples:
            - "2026-06-20T09:00:00+07:00"
        registration_deadline:
          type: string
          format: date-time
          description: "Batas waktu pendaftaran baru (opsional). Harus sebelum atau sama dengan `date`."
          examples:
            - "2026-06-15T23:59:59+07:00"
        quota:
          type: integer
          minimum: 1
          description: "Kuota baru (opsional). Tidak boleh lebih kecil dari jumlah peserta aktif yang sudah terdaftar."
          examples:
            - 75
    EventResponse:
      type: object
      description: "Detail lengkap sebuah event."
      properties:
        id:
          type: integer
          description: "ID unik event."
          examples:
            - 42
        title:
          type: string
          description: "Judul event."
          examples:
            - "Workshop Python Lanjutan"
        description:
          type: string
          description: "Deskripsi lengkap event."
          examples:
            - "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
        date:
          type: string
          format: date-time
          description: "Tanggal dan waktu pelaksanaan."
          examples:
            - "2026-06-15T09:00:00"
        registration_deadline:
          type: string
          format: date-time
          description: "Batas waktu pendaftaran."
          examples:
            - "2026-06-10T23:59:59"
        quota:
          type: integer
          description: "Kapasitas maksimum peserta."
          examples:
            - 50
        status:
          type: string
          enum: [active, deleted]
          description: "Status event. `\"active\"` berarti event dapat dilihat dan didaftarkan; `\"deleted\"` berarti event telah dihapus oleh admin."
          examples:
            - "active"
        created_at:
          type: string
          format: date-time
          description: "Waktu event dibuat di sistem."
          examples:
            - "2026-03-01T08:00:00"
        registration_closed:
          type: boolean
          description: "Bernilai `true` jika batas waktu pendaftaran sudah lewat (dihitung saat respons dikirim). Ini adalah field kalkulasi, bukan disimpan di database."
          examples:
            - false

    # ------------------------------------------------------------------ Registrations
...
    Page_EventResponse:
      type: object
      description: "Halaman berisi daftar event, beserta informasi pagination."
      properties:
        items:
          type: array
          description: "Daftar event pada halaman ini."
          items:
            $ref: "#/components/schemas/EventResponse"
        total_items:
          type: integer
          description: "Total event di semua halaman."
          examples:
            - 85
        page:
          type: integer
          description: "Nomor halaman saat ini (dimulai dari 1)."
          examples:
            - 1
        page_size:
          type: integer
          description: "Jumlah item yang diminta per halaman."
          examples:
            - 20
        total_pages:
          type: integer
          description: "Total jumlah halaman yang tersedia (dihitung dari `total_items / page_size`)."
          examples:
            - 5

    # ------------------------------------------------------------------ Errors
    ErrorDetail:
      type: object
      description: "Detail error yang dikembalikan oleh API."
      required: [code, message, httpStatus]
      properties:
        code:
          type: string
          description: "Kode error yang unik dan dapat dibaca program. Gunakan nilai ini untuk penanganan error di sisi klien."
          examples:
            - "UNAUTHORIZED"
        message:
          type: string
          description: "Pesan error yang dapat dibaca manusia, menjelaskan apa yang terjadi."
          examples:
            - "Could not validate credentials."
        httpStatus:
          type: integer
          description: "Kode HTTP status, sama dengan status code respons."
          examples:
            - 401
    ErrorEnvelope:
      type: object
      description: "Struktur JSON standar untuk semua respons error di API-X."
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      examples:
        unauthorized:
          summary: "401 Unauthorized"
          value:
            error:
              code: "UNAUTHORIZED"
              message: "Could not validate credentials."
              httpStatus: 401
        forbidden:
          summary: "403 Forbidden"
          value:
            error:
              code: "FORBIDDEN"
              message: "You do not have permission to perform this action."
              httpStatus: 403
        not_found:
          summary: "404 Not Found"
          value:
            error:
              code: "EVENT_NOT_FOUND"
              message: "Event not found."
              httpStatus: 404
        validation_error:
          summary: "422 Validation Error"
          value:
            error:
              code: "VALIDATION_ERROR"
              message: "value is not a valid email address"
              httpStatus: 422
...
paths:
...
  /admin/events:
    post:
      tags: [admin-events]
      summary: "Buat event baru (khusus admin)"
      description: |
        Membuat event baru dengan informasi judul, deskripsi, tanggal, batas pendaftaran, dan kuota.

        **Akses**: Khusus admin — memerlukan JWT token dengan role `admin`.

        **Aturan**:
        - `registration_deadline` harus sebelum atau sama dengan `date`.
        - `quota` minimal 1.
        - Event yang baru dibuat langsung berstatus `active`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreate"
            example:
              title: "Workshop Python Lanjutan"
              description: "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
              date: "2026-06-15T09:00:00+07:00"
              registration_deadline: "2026-06-10T23:59:59+07:00"
              quota: 50
      responses:
        "201":
          description: "Event berhasil dibuat."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
              example:
                id: 42
                title: "Workshop Python Lanjutan"
                description: "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
                date: "2026-06-15T09:00:00"
                registration_deadline: "2026-06-10T23:59:59"
                quota: 50
                status: "active"
                created_at: "2026-03-09T08:00:00"
                registration_closed: false
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "403":
          description: "Token valid tapi bukan role admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "You do not have permission to perform this action."
                  httpStatus: 403
        "422":
          description: "Data tidak valid — misalnya `registration_deadline` setelah `date`, atau `quota` kurang dari 1."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalid_date_range:
                  summary: "registration_deadline setelah date"
                  value:
                    error:
                      code: "INVALID_DATE_RANGE"
                      message: "registration_deadline must be on or before the event date"
                      httpStatus: 422
                validation_error:
                  summary: "Field tidak valid"
                  value:
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Input should be greater than or equal to 1"
                      httpStatus: 422
  /admin/events/{event_id}:
    get:
      tags: [admin-events]
      summary: "Ambil detail event (khusus admin)"
      description: |
        Mengembalikan detail lengkap satu event berdasarkan ID, termasuk event yang sudah dihapus.

        **Akses**: Khusus admin — memerlukan JWT token dengan role `admin`.

        Berbeda dengan `GET /events/{event_id}` (publik) yang hanya mengembalikan event aktif.
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: "ID unik event yang ingin dilihat detailnya."
          schema:
            type: integer
            minimum: 1
          examples:
            valid:
              value: 42
      responses:
        "200":
          description: "Detail event berhasil diambil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
              example:
                id: 42
                title: "Workshop Python Lanjutan"
                description: "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
                date: "2026-06-15T09:00:00"
                registration_deadline: "2026-06-10T23:59:59"
                quota: 50
                status: "active"
                created_at: "2026-03-09T08:00:00"
                registration_closed: false
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "403":
          description: "Token valid tapi bukan role admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "You do not have permission to perform this action."
                  httpStatus: 403
        "404":
          description: "Event dengan ID tersebut tidak ditemukan."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "EVENT_NOT_FOUND"
                  message: "Event not found."
                  httpStatus: 404

    put:
      tags: [admin-events]
      summary: "Perbarui informasi event (khusus admin)"
      description: |
        Memperbarui satu atau lebih field event yang sudah ada. Hanya field yang disertakan
        dalam request body yang akan diperbarui (partial update).

        **Akses**: Khusus admin — memerlukan JWT token dengan role `admin`.

        **Aturan**:
        - Jika `quota` diperbarui, nilai baru tidak boleh lebih kecil dari jumlah peserta aktif saat ini.
        - Jika `registration_deadline` dan `date` keduanya disertakan, `registration_deadline` harus ≤ `date`.
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: "ID unik event yang akan diperbarui."
          schema:
            type: integer
            minimum: 1
          examples:
            valid:
              value: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
            example:
              title: "Workshop Python Lanjutan — Edisi Revisi"
              quota: 75
      responses:
        "200":
          description: "Event berhasil diperbarui."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
              example:
                id: 42
                title: "Workshop Python Lanjutan — Edisi Revisi"
                description: "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
                date: "2026-06-15T09:00:00"
                registration_deadline: "2026-06-10T23:59:59"
                quota: 75
                status: "active"
                created_at: "2026-03-09T08:00:00"
                registration_closed: false
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "403":
          description: "Token valid tapi bukan role admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "You do not have permission to perform this action."
                  httpStatus: 403
        "404":
          description: "Event dengan ID tersebut tidak ditemukan."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "EVENT_NOT_FOUND"
                  message: "Event not found."
                  httpStatus: 404
        "409":
          description: "Kuota baru lebih kecil dari jumlah peserta aktif."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "QUOTA_BELOW_PARTICIPANTS"
                  message: "New quota is below the current number of active participants."
                  httpStatus: 409
        "422":
          description: "Data tidak valid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Input should be greater than or equal to 1"
                  httpStatus: 422

    delete:
      tags: [admin-events]
      summary: "Hapus event (khusus admin)"
      description: |
        Menghapus event secara soft delete — status event diubah menjadi `deleted` dan event
        tidak lagi muncul di endpoint publik, tapi data tetap tersimpan di database.

        **Akses**: Khusus admin — memerlukan JWT token dengan role `admin`.

        Mengembalikan **204 No Content** jika berhasil (tidak ada body respons).
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: "ID unik event yang akan dihapus."
          schema:
            type: integer
            minimum: 1
          examples:
            valid:
              value: 42
      responses:
        "204":
          description: "Event berhasil dihapus. Tidak ada body respons."
        "401":
          description: "Token tidak ada, tidak valid, atau sudah kedaluwarsa."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Could not validate credentials."
                  httpStatus: 401
        "403":
          description: "Token valid tapi bukan role admin."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "You do not have permission to perform this action."
                  httpStatus: 403
        "404":
          description: "Event dengan ID tersebut tidak ditemukan."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "EVENT_NOT_FOUND"
                  message: "Event not found."
                  httpStatus: 404

  # ================================================================ Public Events
  /events:
    get:
      tags: [events]
      summary: "Daftar event yang tersedia"
      description: |
        Mengembalikan daftar event yang berstatus `active` secara terpaginasi.

        **Akses**: Publik — tidak memerlukan autentikasi.

        Hasil diurutkan berdasarkan tanggal terdekat. Event dengan status `deleted` tidak ditampilkan.
      parameters:
        - name: page
          in: query
          required: false
          description: "Nomor halaman yang diminta (dimulai dari 1)."
          schema:
            type: integer
            minimum: 1
            default: 1
          examples:
            first_page:
              summary: "Halaman pertama"
              value: 1
        - name: page_size
          in: query
          required: false
          description: "Jumlah event per halaman (minimal 1, maksimal 100)."
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          examples:
            default_size:
              summary: "20 per halaman (default)"
              value: 20
      responses:
        "200":
          description: "Daftar event berhasil diambil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_EventResponse"
              example:
                items:
                  - id: 42
                    title: "Workshop Python Lanjutan"
                    description: "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
                    date: "2026-06-15T09:00:00"
                    registration_deadline: "2026-06-10T23:59:59"
                    quota: 50
                    status: "active"
                    created_at: "2026-03-09T08:00:00"
                    registration_closed: false
                total_items: 85
                page: 1
                page_size: 20
                total_pages: 5
        "422":
          description: "Parameter tidak valid (misalnya `page` kurang dari 1 atau `page_size` lebih dari 100)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Input should be greater than or equal to 1"
                  httpStatus: 422
  /events/{event_id}:
    get:
      tags: [events]
      summary: "Ambil detail satu event"
      description: |
        Mengembalikan detail lengkap satu event berdasarkan ID.

        **Akses**: Publik — tidak memerlukan autentikasi.

        Hanya mengembalikan event dengan status `active`. Event yang sudah dihapus
        akan mengembalikan 404, bukan data event.
      parameters:
        - name: event_id
          in: path
          required: true
          description: "ID unik event yang ingin dilihat detailnya."
          schema:
            type: integer
            minimum: 1
          examples:
            valid:
              value: 42
      responses:
        "200":
          description: "Detail event berhasil diambil."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
              example:
                id: 42
                title: "Workshop Python Lanjutan"
                description: "Workshop intensif satu hari tentang FastAPI dan Pydantic v2."
                date: "2026-06-15T09:00:00"
                registration_deadline: "2026-06-10T23:59:59"
                quota: 50
                status: "active"
                created_at: "2026-03-09T08:00:00"
                registration_closed: false
        "404":
          description: "Event tidak ditemukan atau sudah dihapus."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: "EVENT_NOT_FOUND"
                  message: "Event not found."
                  httpStatus: 404

  # ================================================================ Registrations
...